made mediapipe model. Found frame mistake in app

+++ computation/utils.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 def get_coords(landmarks: NormalizedLandmarkList, init_image_shape: Tuple[int, int], points: list) -> np.array:
     image_hight, image_width = init_image_shape
-    return np.array([(landmarks.landmark[idx].x * image_width, landmarks.landmark[idx].y * image_hight) for idx in  points])
+    return np.array([(landmarks.landmark[idx].x * image_width, landmarks.landmark[idx].y * image_hight) for idx in  points]).astype(int)
 
 
 def calculate_geom_center(point_group: np.array):
